Style section 1 of landing page
Saving as css code seems to be stable enough to
move onto section 2. The only confusion I have at
the moment is on whether the content in section 1
should be centered or grow with the header. It is
hard to tell from the png TOP provided so I am
moving on and can change it afterwards if needed.